requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

argon2_cffi==23.1.0
fastapi==0.115.5
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.9.2
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2

